DB/Creature: Partially revert 72802f54287acacec1f98057807075ec6e92562c
If path_id is loaded from c_t_a and not from c_a, creature requires MovementType = 2 in c_t. Otherwise waypoint will be not loaded. Reverting instead of changing MovementType since MovementType in c_t with huge chance is not really what we need, that column before that commit was barely used. And waypoints better to store in c_a and not in c_t_a since it's easier to interract with them.