DB/SAI: Remove wrongly used SmartActionSummonCreatureFlags
Closes #26595

After 523971f action_param5 is used as SmartActionSummonCreatureFlags and 1 is PersonalSpawn. And somehow currently action_param5 = 1 is used in 40 actions even if that param wasn't used by core before 523971f. Because of that those summoned creatures became personal spawns of creatures or objects and were not visible to player as they're supposed to be