DB/SAI: Get rid of AddAura pt 5
It will be removed after 2 weeks for several reasons. Some of them: It was heavily abused and creates not obvious and thus dangerous logic in scripts, making them easily breakable. In lots of cases was used as hacky version of Cross Cast(one creature applies aura on another creature) while it should be casted by creature and not applied to it by another creature. In some cases it was used to ensure aura will be applied always (something like casting a spell with triggered cast flag). In nearly all other cases used for no reason. There's no such action in Blizz's scripting system, the reason why some auras are applied without SPELL_START and SPELL_GO has to be different. On top of that some auras in addons are channeling spells, some has effect Spawn, some has on-cast visuals. Using AddAura breaks them. That means if Blizz indeed used some kind of system only for on-spawn spells, all spells are casted or some are handled in another place.